Modularize views.py
Put alogirthm in it's own seperate file algorithm.py, created indexView
that links to search.html, and explore now links to explore.html.
