+++ django/templates/explore.html
@@ -37,9 +37,6 @@
     data: {},
     success:function(data)
     {
-    alert("Movies loaded");
-    alert(data)
-    alert(data[0].fields)
     var test = data;
     var text;
