+++ django/templates/search.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+<form>
+  <br>
+  <input type="text" name="Choice"><br>
+</form>
+</body>
+</html>
